pyproject.toml
[project]
name = "openrouter-exercise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
]
